restricted pages

+++ base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from django.db.models import Q
 from .models import Room
@@ -14,6 +16,9 @@ from .models import Topic
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -33,9 +38,11 @@ def loginPage(request):
 
     context = {}
     return render(request,'base/login_register.html', context)
+
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
 
@@ -54,6 +61,7 @@ def room(request, pk):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,9 +72,13 @@ def createRoom(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse('Restricted: Unmatched roles')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -76,8 +88,12 @@ def updateRoom(request, pk):
 
     context = {'form' : form}
     return render(request,'base/room_form.html', context)
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host :
+        return HttpResponse('Restricted: Unmatched roles')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
